Update WindowsDX template
