--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/BlankApp.Windows/.template.config/dotnetcli.host.json
@@ -1,3 +1,0 @@
-{
-
-}
