--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/BlankApp.Windows/.template.config/template.json
@@ -1,20 +1,0 @@
-{
-    "author": "MonoGameTeam",
-    "classifications": [ "MonoGame" ],
-    "name": "MonoGame Windows Application",
-    "groupIdentity": "MonoGame.BlankApp.Windows",
-    "identity": "MonoGame.BlankApp.Windows.CSharp",
-    "shortName": "mgwindows",
-    "tags": {
-        "language": "C#",
-        "type": "project"
-    },
-    "primaryOutputs": [ 
-                        {
-                            "path": "BlankApp.Windows.csproj"
-                        },
-                      ], 
-    "sourceName": "BlankApp", 
-    "preferNameDirectory" : "true",
-    "description": "A MonoGame game project for Windows using DirectX."
-}
