--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/BlankApp.Windows/Content/Content.mgcb
@@ -1,15 +1,0 @@
-
-#----------------------------- Global Properties ----------------------------#
-
-/outputDir:bin/$(Platform)
-/intermediateDir:obj/$(Platform)
-/platform:Windows
-/config:
-/profile:Reach
-/compress:False
-
-#-------------------------------- References --------------------------------#
-
-
-#---------------------------------- Content ---------------------------------#
-
